Utility functions and convenience functions consolidated and placed to LintUtils.kt
