+++ lint/src/main/kotlin/kiwi/orbit/compose/lint/detectors/MaterialDialogWithOrbitButtonsDetector.kt
@@ -8,15 +8,10 @@ import com.android.tools.lint.detector.api.JavaContext
 import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
-import com.android.tools.lint.detector.api.computeKotlinArgumentMapping
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiJavaFile
-import com.intellij.psi.PsiMember
 import com.intellij.psi.PsiMethod
 import com.intellij.psi.PsiNamedElement
 import org.jetbrains.uast.UBlockExpression
 import org.jetbrains.uast.UCallExpression
-import org.jetbrains.uast.ULambdaExpression
 import org.jetbrains.uast.skipParenthesizedExprDown
 import org.jetbrains.uast.tryResolveNamed
 
@@ -51,7 +46,7 @@ class MaterialDialogWithOrbitButtonsDetector : Detector(), SourceCodeScanner {
 
     override fun visitMethodCall(context: JavaContext, node: UCallExpression, method: PsiMethod) {
         if (method.isInPackageName(PACKAGE_MATERIAL_ALERT)) {
-            val confirmButtonArgument = getArgument(node, method, MATERIAL_ALERT_CONFIRM_BUTTON_PARAM)
+            val confirmButtonArgument = method.getArgument(node, MATERIAL_ALERT_CONFIRM_BUTTON_PARAM)
 
             (confirmButtonArgument?.body as? UBlockExpression)?.let { body ->
                 val resolvedBodyExpression = body.resolveFirstBodyExpression() ?: return@let
@@ -77,7 +72,7 @@ class MaterialDialogWithOrbitButtonsDetector : Detector(), SourceCodeScanner {
                 }
             }
 
-            val dismissButtonArgument = getArgument(node, method, MATERIAL_ALERT_DISMISS_BUTTON_PARAM)
+            val dismissButtonArgument = method.getArgument(node, MATERIAL_ALERT_DISMISS_BUTTON_PARAM)
 
             (dismissButtonArgument?.body as? UBlockExpression)?.let { body ->
                 val resolvedBodyExpression = body.resolveFirstBodyExpression() ?: return@let
@@ -103,29 +98,6 @@ class MaterialDialogWithOrbitButtonsDetector : Detector(), SourceCodeScanner {
                 }
             }
         }
-    }
-
-    private fun getArgument(
-        node: UCallExpression,
-        method: PsiMethod,
-        argumentName: String,
-    ): ULambdaExpression? = computeKotlinArgumentMapping(node, method)
-        .orEmpty()
-        .filter { (_, parameter) ->
-            parameter.name == argumentName
-        }
-        .keys
-        .filterIsInstance<ULambdaExpression>()
-        .firstOrNull()
-
-    private fun PsiMethod.isInPackageName(packageName: String): Boolean {
-        val actual = (containingFile as? PsiJavaFile)?.packageName
-        return packageName == actual
-    }
-
-    private fun PsiElement.getPackageName(): String? = when (this) {
-        is PsiMember -> this.containingClass?.qualifiedName?.let { it.substring(0, it.lastIndexOf(".")) }
-        else -> null
     }
 
     private fun UBlockExpression.resolveFirstBodyExpression(): PsiNamedElement? {
